Update ConfigManager and related classes
