+++ src/com/projectkorra/ProjectKorra/ProjectKorra.java
@@ -16,6 +16,7 @@ import com.projectkorra.ProjectKorra.Utilities.Updater;
 import com.projectkorra.ProjectKorra.airbending.AirbendingManager;
 import com.projectkorra.ProjectKorra.chiblocking.ChiComboManager;
 import com.projectkorra.ProjectKorra.chiblocking.ChiblockingManager;
+import com.projectkorra.ProjectKorra.configuration.ConfigManager;
 import com.projectkorra.ProjectKorra.earthbending.EarthbendingManager;
 import com.projectkorra.ProjectKorra.firebending.FirebendingManager;
 import com.projectkorra.ProjectKorra.waterbending.WaterbendingManager;
@@ -25,7 +26,6 @@ public class ProjectKorra extends JavaPlugin {
 	public static long time_step = 1;
 	public static ProjectKorra plugin;
 	public static Logger log;
-    public static Config deathMsgConfig;
 
 	public Updater updater;
 	
@@ -33,9 +33,7 @@ public class ProjectKorra extends JavaPlugin {
 	public void onEnable() {
 		ProjectKorra.log = this.getLogger();
 		plugin = this;
-        deathMsgConfig = new Config(new File("deathmessages.yml"));
 		new ConfigManager(this);
-        new DeathMessageConfigManager(this);
 		new GeneralMethods(this);
 		updater = new Updater(this, "http://projectkorra.com/forum/forums/dev-builds.16/index.rss");
 		new Commands(this);
@@ -45,8 +43,6 @@ public class ProjectKorra extends JavaPlugin {
 		new ComboModuleManager();
 		new ComboManager();
 		new ChiComboManager();
-
-		ConfigManager.configCheck();
 
 		DBConnection.host = getConfig().getString("Storage.MySQL.host");
 		DBConnection.port = getConfig().getInt("Storage.MySQL.port");
